backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Config & typing
pydantic-settings==2.2.1
python-dotenv==1.0.1
loguru==0.7.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Async HTTP client
httpx==0.27.0

# Database (will be used in Phase 2)
sqlmodel==0.0.16
asyncpg==0.29.0
aiosqlite==0.20.0
greenlet==3.0.3

# Background jobs (Phase 3+)
celery==5.4.0
redis==5.0.4

# Google APIs (Phase 2+)
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.126.0

# AI Agent (Phase 4)
google-adk

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
httpx[cli]==0.27.0

# --- Phase 2 additions ---
alembic==1.13.2
cryptography>=42.0.4          # for Fernet encryption
python-multipart==0.0.9       # FastAPI redirection helper
